Added isAdminOfSpace and isSuperAdminOfDoDAO fields to jwt token
