+++ academy-ui/src/app/api/auth/[...nextauth]/authOptions.ts
@@ -1,10 +1,10 @@
-import { isUserAdminOfSpace } from '@/app/api/helpers/auth/checkEditSpacePermission';
-import { isDoDAOSuperAdmin } from '@/app/api/helpers/auth/isSuperAdmin';
+import { isUserAdminOfSpace } from '@/app/api/helpers/space/checkEditSpacePermission';
+import { isDoDAOSuperAdmin } from '@/app/api/helpers/space/isSuperAdmin';
 import { getAuthOptions } from '@dodao/web-core/api/auth/authOptions';
 import { authorizeCrypto } from '@dodao/web-core/api/auth/authorizeCrypto';
 import { User } from '@dodao/web-core/types/auth/User';
 import { PrismaAdapter } from '@next-auth/prisma-adapter';
-import { PrismaClient } from '@prisma/client';
+import { PrismaClient, Space } from '@prisma/client';
 import { AuthOptions } from 'next-auth';
 
 const p = new PrismaClient();
@@ -41,12 +41,16 @@ export const authOptions: AuthOptions = getAuthOptions(
             where: { id: token.sub },
           });
 
+          const space: Space | null = await p.space.findUnique({
+            where: { id: dbUser?.spaceId },
+          });
+
           if (dbUser) {
             token.spaceId = dbUser.spaceId;
             token.username = dbUser.username;
             token.authProvider = dbUser.authProvider;
             token.accountId = dbUser.id;
-            token.isAdminOfSpace = isUserAdminOfSpace(dbUser.username);
+            token.isAdminOfSpace = isUserAdminOfSpace(dbUser.username, space!);
             token.isSuperAdminOfDoDAO = isDoDAOSuperAdmin(dbUser.username);
           }
         }
