+++ academy-ui/src/app/api/helpers/space/isSuperAdmin.ts
@@ -1,15 +1,13 @@
-import { DoDaoJwtTokenPayload } from '@dodao/web-core/types/auth/Session';
 import { getDecodedJwtFromContext } from '@/app/api/helpers/permissions/getJwtFromContext';
 import { NextRequest } from 'next/server';
 
 const superAdminsFromEnv = process.env.DODAO_SUPERADMINS ? process.env.DODAO_SUPERADMINS.split(',') : [];
 const superAdminsArray = [...superAdminsFromEnv, '0x470579d16401a36BF63b1428eaA7189FBdE5Fee9', 'robinnagpal.tiet@gmail.com'].map((u) => u.toLowerCase().trim());
-export async function isDoDAOSuperAdmin(context: NextRequest): Promise<DoDaoJwtTokenPayload | null> {
-  const decoded = await getDecodedJwtFromContext(context);
-  if (superAdminsArray.includes(decoded!.username.toLowerCase())) {
-    return decoded;
+export function isDoDAOSuperAdmin(username: string): boolean {
+  if (superAdminsArray.includes(username.toLowerCase())) {
+    return true;
   }
-  return null;
+  return false;
 }
 
 export function isSuperAdminOfDoDAO(username: string): boolean {
@@ -18,7 +16,7 @@ export function isSuperAdminOfDoDAO(username: string): boolean {
 
 export async function validateSuperAdmin(context: NextRequest) {
   const decoded = await getDecodedJwtFromContext(context);
-  if (!isDoDAOSuperAdmin(context)) {
+  if (!isDoDAOSuperAdmin(decoded!.username)) {
     throw new Error(`Not authorized ${decoded?.username}`);
   }
 }
